broken migrate_db into individual schema version files
